fix: All 6 priorities - bcrypt pin, auto-seed, itinerary tests, 403->404, remove stubs, Numeric expense, FK pragma

+++ travel-planner/backend/app/models/expense.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Float
+from sqlalchemy import Column, Integer, String, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from app.database.base import Base
 
@@ -7,7 +7,7 @@ class Expense(Base):
 
     id = Column(Integer, primary_key=True)
     trip_id = Column(Integer, ForeignKey("trips.id", ondelete="CASCADE"), nullable=False, index=True)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(10, 2), nullable=False)
     category = Column(String, nullable=False)
     description = Column(String, nullable=True)
